fix bug 13669
static const GLchar* verification_snippet =
                "    vec4 goku_color = imageLoad(goku, ivec2(0,0));\n"
                "\n"
                "    imageStore(goku, ivec2(0,0), vec4(0, 1, 0, 1));\n"
                "\n"
                "    if (vec4(1, 0, 0, 0) != goku_color)\n"
                "    {\n"
                "        result = vec4(1, 0, 0, 1);\n"
                "    }\n";

If vertex shader read in correct value, it will store the green color to the same location so later when TES, TCS, GS, FS try to read it, it got the bad color (green color) and generate failure color to the result.

If want to test the imageStore save the correct green color to the location, I think the read location and write location shall be different. 

Luckily, the texture is a 2D array. It has a lot of space to read from.